LeBonPain Prototype base project.
This is an alpha version of a prototype distributed software, where clients of a large bakery (Le Bon Pain) can place orders for baked goods that will be delivered at scheduled times. This is important for cafes, restaurants, grocers, and other vendors who rely on a constant supply of fresh baked products.

The LeBonPain prototype is based on the distributed architecture, where customers install a local version of the software and then connect to the LeBonPain servers using a secured and encrypted TCP connection. Orders can then be placed, using registered accounts and typical payment gateways such as credit card and Paypal.